requirements.txt
autopep8==1.5.7
blis==0.7.4
catalogue==2.0.6
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
cupy-cuda111==9.5.0
cycler==0.10.0
cymem==2.0.5
datasets==1.9.0
dill==0.3.4
#en-core-web-sm==3.1.0
fastrlock==0.6
filelock==3.2.0
Flask==2.0.1
fsspec==2021.9.0
huggingface-hub==0.0.17
idna==3.2
install==1.3.4
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.2
llvmlite==0.37.0
MarkupSafe==2.0.1
matplotlib==3.4.3
multiprocess==0.70.12.2
murmurhash==1.0.5
numba==0.54.0
numpy==1.19.5
olefile==0.46
packaging==21.0
pandas==1.3.3
pathy==0.6.0
Pillow==8.3.1
pip==21.3
preshed==3.0.5
pyarrow==5.0.0
pycodestyle==2.7.0
pydantic==1.8.2
pynndescent==0.5.4
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
regex==2021.9.24
requests==2.26.0
sacremoses==0.0.46
scikit-learn==1.0
scipy==1.7.1
setuptools==58.2.0
six==1.16.0
smart-open==5.2.1
spacy==3.1.3
spacy-legacy==3.0.8
srsly==2.4.1
thinc==8.0.11
threadpoolctl==2.2.0
tokenizers==0.10.3
toml==0.10.2
torch==1.9.1
torchaudio==0.9.1
torchvision==0.10.1
tqdm==4.62.3
transformers==4.11.1
typer==0.4.0
typing-extensions==3.10.0.2
umap-learn==0.5.1
urllib3==1.26.7
wasabi==0.8.2
Werkzeug==2.0.1
wheel==0.37.0
wincertstore==0.2
xxhash==2.0.2
